rework DigitSet's all_sets and iter_all
Rather than a custom iterator type that yields elements in an
unspecified order, lazily construct a sorted list of all the digit sets
and return it as a slice. Use std's iterator adapters to iterate through
it (getting ExactSizeIterator, DoubleEndedIterator, and TrustedLen
optimizations for free).

This is a semver-breaking change.