update to rustyline v11, disable its default features
Editor is now also generic over a History type, but there's a convenient
DefaultEditor type, removing the somewhat ugly type annotations that
were previously required.

Also disable the default features (custom-bindings, with-dirs,
with-file-history) because we're not using any of them and it forces
DefaultHistory (and therefore DefaultEditor) to be memory-only with no
file handing code compiled in.

Warning: trying to save the history will *panic* with unimplemented!()
rather than just returning an error. Good thing I don't call that
method.